adds player slot in text at the top and change movement

+++ src/apps/snacman/snacman/simulations/snacgame/component/Spawner.h
@@ -9,7 +9,7 @@ namespace component {
 
 struct Spawner
 {
-    math::Position<2, int> mSpawnPosition;
+    math::Position<2, float> mSpawnPosition;
 };
